Address @pawelbaran's comments and move namespace
