+++ Structure_oM/Reinforcement/PanelRebarIntent.cs
@@ -20,21 +20,26 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using BH.oM.Analytical.Elements;
 using BH.oM.Base;
-using BH.oM.Geometry;
+using BH.oM.Quantities.Attributes;
+using System.Collections.Generic;
 using System.ComponentModel;
 
-namespace BH.oM.Structure.SurfaceProperties.Reinforcement
+
+namespace BH.oM.Structure.Reinforcement
 {
-    [Description("A region defining the area of a Panel to be reinforced. The region must exist within the ExternalEdges of the Panel.")]
-    public class ReinforcementRegion : BHoMObject, IRegion
+    [Description("A collection of PanelReinforcement objects that represent the rebar intent. \n" +
+        "This is stored as layers of reinforcement amongst regions within the Panel.")]
+    public class PanelRebarIntent : BHoMObject, IFragment, IRebarIntent
     {
-        [Description("Local x, y, and z axes of the region as a vector Basis. Defaults to world axes.")]
-        public virtual Basis Orientation { get; set; } = null;
+        /***************************************************/
+        /**** Properties                                ****/
+        /***************************************************/
 
-        [Description("A 2D curve defining the external boundaries of the region to be reinforced. The curve must exist within the ExternalEdges of the Panel.")]
-        public virtual ICurve Perimeter { get; set; } = null;
+        [Description("A collection of PanelReinforcement objects.")]
+        public virtual List<PanelReinforcement> PanelReinforcement { get; set; } = new List<PanelReinforcement>();
+
+        /***************************************************/
     }
-
 }
+
